+++ vm/CHANGELOG.md
@@ -7,11 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `PUBLIC_SENDER` available to session [#3341]
+
 ### Changed
 
 - Change `execution` module to use `execution::Config` [#3437]
+- Change `dusk-core` dependency to `1.0.1-alpha` [#3341]
+- Change `piecrust` dependency to `0.27.1` [#3341]
 
-[1.0.0] - 2025-01-23
+## [1.0.0] - 2025-01-23
 
 ### Changed
 
@@ -25,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- Issues -->
 [#3235]: https://github.com/dusk-network/rusk/issues/3235
+[#3341]: https://github.com/dusk-network/rusk/issues/3341
 [#3405]: https://github.com/dusk-network/rusk/issues/3405
 [#3437]: https://github.com/dusk-network/rusk/issues/3437